+ using CuArrays & Sys.iswindows for MPI finalize
The 1D kernels depend on CuArrays, but do not
specify using CuArrays in the file, so they fail
to run.

Also, the MPI.finalize_atexit() is not supported
on windows, and must be used conditionally. A
Sys.iswindows() was used to conditionally use
MPI.finalize_atexit().

With this PR, the 1D examples run without error.